Disable abilities extension-wide
Due to the way that abilities are currently implemented, load on the
Daylight API side is very high. We don't have the cycles right now to
implement the underlying changes that would improve API load, so for now
we disable abilities fetching, as well as access to abilities in the UI.

The code that is commented out is targeted to only remove the parts that
link to the abilities UI and the parts that call out to the Daylight
API; the rest remains in place so that turning things back on once the
fetching strategy can be improved is straightforward.