Fix Ledger message signing on built-in non-ETH/MATIC networks
Ledger was checking specifically for those two networks from a constant
that was never updated. We switch instead to check whether the network
in question shares a derivation path with the one used by the Eth app
(aka the Ethereum derivation path).

Custom networks added by users currently don't set a derivation path, so
they will still fail this test (as well as the guards on signing
messages and typed messages, which are already using derivation paths).