Undo a change to how typed data sig request chain ids are handled
The change forced the chain id of a typed data signature request to that
of the current change. In fact, EIP712 specifies that if the chain id is
different from the current one, the user agent should *reject* the
signature rather than adjust it and produce a valid signature for a
different chain than requested.

The SigningService already handles the rejection behavior, so the code
adjusting the chain id is removed without making any additional
adjustments.