Drop Wallet Connect code
This doesn't work, and also makes it so the background package is now
depending on the window-provider package, which creates a situation
where manifest v3 service worker code would try to use the `window`
object, which is no longer available.

The Wallet Connect work itself was never enabled, and is based on Wallet
Connect v1 whose backend services have been turned off for some time.
This commit can always be reverted or mined for architecture in the
future to implement a Wallet Connect v2-compatible approach.