Bump webext-redux and other packages and update manifest to v3
The temporary change done here is commenting out any localStorage usage
in the background script as it is unavailable in a service worker
context. Currently the relevant calls just become noops (for setters) or
return blank values (for getters). These should be updated to use
extension-local storage (and asyncified accordingly).

Note that the patch for webext-redux does not work for version 3.0 that
is used here. The patch needs to be updated accordingly, and likely some
commit massaging is in order after.