Properly guard `window` references in shared modules
The service worker context doesn't have a `window` variable. Most of
these references were already guarding for an `undefined` value, but not
for the variable not to exist at all, which means they were hitting a
runtime error. We now do the full `typeof === "undefined"` check to
ensure we don't fail in these cases.