Fix snackbar assertions
These were buggy due to the duplicate transition elements in the popup.
Additionally assertions for snackbars that don't appear immediately resolve
as passed.

Another issue is that even though it's technically not visible, elements with 0
opacity are still considered visible by playwright. This adds a testid to a
span wrapper inside the snackbar, so the `toBeVisible` matcher works.